Optimized transition between 4-digit srcreen and menu view.
The large slideshow image on front screen load not all at once. First load the first 2 - then load the next 2 1 sec later, and finally the last 2 another 1sec later. This was there is no choppiness during the transition on android phones.

This is kind of jacky. Might have to come up with a better way later